chatbot-backend/requirements.txt
Flask==3.0.3
google-generativeai==0.7.2
python-dotenv==1.0.1
requests==2.32.3
langdetect==1.0.9
deep-translator==1.11.4
rapidfuzz==3.14.0
flask-cors
